change
There are a few grammar errors in the bulleted list part of your About Me section:
programing needs to be changed to its correct spelling -> programming.
You have a space character between the words visual basics and the comma separating the words in this sequential list. Try removing this space character.
Gradation needs to be changed to its correct spelling of -> Graduation.